requirements.txt
Django==4.2.13
djangorestframework==3.14.0
channels==4.1.0
channels-redis==4.2.0
Pillow==10.3.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
crispy-forms
crispy-bootstrap4
factory_boy==3.3.0
pytest-django==4.8.0